+++ server/compaction-coordinator/src/test/java/org/apache/accumulo/coordinator/QueueSummariesTest.java
@@ -36,7 +36,7 @@ public class QueueSummariesTest {
     return new TServerInstance(tserver + ":9997", 0);
   }
 
-  private PrioTserver npt(String tserver, long prio) {
+  private PrioTserver npt(String tserver, short prio) {
     return new PrioTserver(ntsi(tserver), prio);
   }
 
@@ -47,7 +47,7 @@ public class QueueSummariesTest {
     List<TCompactionQueueSummary> summaries = new ArrayList<>();
 
     for (int i = 0; i < data.length; i += 2) {
-      summaries.add(new TCompactionQueueSummary(data[i], Long.parseLong(data[i + 1])));
+      summaries.add(new TCompactionQueueSummary(data[i], Short.parseShort(data[i + 1])));
     }
 
     queueSum.update(tsi, summaries);
@@ -62,33 +62,33 @@ public class QueueSummariesTest {
     update(queueSum, "ts3", "q1", "5", "q2", "5");
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts3", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts3", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts3", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts3", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q3"));
     }
 
-    queueSum.removeSummary(ntsi("ts2"), "q1", 5);
+    queueSum.removeSummary(ntsi("ts2"), "q1", (short) 5);
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts3", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts3", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts3", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts3", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q3"));
     }
 
-    queueSum.removeSummary(ntsi("ts3"), "q2", 5);
-    queueSum.removeSummary(ntsi("ts2"), "q3", 5);
+    queueSum.removeSummary(ntsi("ts3"), "q2", (short) 5);
+    queueSum.removeSummary(ntsi("ts2"), "q3", (short) 5);
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts3", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 4), queueSum.getNextTserver("q3"));
-      assertEquals(npt("ts2", 4), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts3", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 4), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts2", (short) 4), queueSum.getNextTserver("q3"));
     }
 
   }
@@ -101,10 +101,10 @@ public class QueueSummariesTest {
     update(queueSum, "ts2", "q1", "5", "q2", "4", "q3", "5");
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q3"));
     }
 
     // an update from the tserver should remove some existing entries
@@ -112,20 +112,20 @@ public class QueueSummariesTest {
     update(queueSum, "ts2", "q1", "7", "q3", "3", "q4", "5");
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts2", 7), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 6), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 3), queueSum.getNextTserver("q3"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q4"));
+      assertEquals(npt("ts2", (short) 7), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 6), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 3), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q4"));
     }
 
-    queueSum.removeSummary(ntsi("ts2"), "q1", 7);
-    queueSum.removeSummary(ntsi("ts1"), "q2", 6);
+    queueSum.removeSummary(ntsi("ts2"), "q1", (short) 7);
+    queueSum.removeSummary(ntsi("ts1"), "q2", (short) 6);
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 4), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 4), queueSum.getNextTserver("q1"));
       assertNull(queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 3), queueSum.getNextTserver("q3"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q4"));
+      assertEquals(npt("ts2", (short) 3), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q4"));
     }
   }
 
@@ -137,48 +137,48 @@ public class QueueSummariesTest {
     update(queueSum, "ts2", "q1", "5", "q2", "4", "q3", "5");
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts1", 5), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 5), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q3"));
     }
 
     queueSum.remove(Set.of(ntsi("ts1")));
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts2", 4), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q3"));
+      assertEquals(npt("ts2", (short) 4), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q3"));
     }
 
-    queueSum.removeSummary(ntsi("ts2"), "q3", 5);
+    queueSum.removeSummary(ntsi("ts2"), "q3", (short) 5);
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts2", 4), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts2", (short) 4), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
       assertNull(queueSum.getNextTserver("q3"));
     }
 
     update(queueSum, "ts1", "q2", "6", "q1", "3");
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 6), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts2", 5), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 6), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts2", (short) 5), queueSum.getNextTserver("q1"));
       assertNull(queueSum.getNextTserver("q3"));
     }
 
     queueSum.remove(Set.of(ntsi("ts2")));
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 6), queueSum.getNextTserver("q2"));
-      assertEquals(npt("ts1", 3), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 6), queueSum.getNextTserver("q2"));
+      assertEquals(npt("ts1", (short) 3), queueSum.getNextTserver("q1"));
       assertNull(queueSum.getNextTserver("q3"));
     }
 
-    queueSum.removeSummary(ntsi("ts1"), "q2", 6);
+    queueSum.removeSummary(ntsi("ts1"), "q2", (short) 6);
 
     for (int i = 0; i < 3; i++) {
-      assertEquals(npt("ts1", 3), queueSum.getNextTserver("q1"));
+      assertEquals(npt("ts1", (short) 3), queueSum.getNextTserver("q1"));
       assertNull(queueSum.getNextTserver("q2"));
       assertNull(queueSum.getNextTserver("q3"));
     }
